requirements.txt
openai-agents
streamlit
pydantic
python-dotenv
pandas
numpy
matplotlib
networkx
geopy 